pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "a1-trainer"
version = "0.1.0"
description = "OneDesk AI/LLM Middleware & Training Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "sqlalchemy[asyncio]>=2.0.35",
    "asyncpg>=0.30.0",
    "alembic>=1.13.0",
    "redis[hiredis]>=5.1.0",
    "httpx>=0.27.0",
    "litellm>=1.50.0",
    "tiktoken>=0.8.0",
    "arq>=0.26.0",
    "pyyaml>=6.0.0",
    "python-multipart>=0.0.12",
    "websockets>=13.0",
    "opentelemetry-api>=1.27.0",
    "opentelemetry-sdk>=1.27.0",
    "opentelemetry-instrumentation-fastapi>=0.48b0",
    "opentelemetry-exporter-otlp>=1.27.0",
    "gptcache>=0.1.44",
]

[project.optional-dependencies]
training = [
    "unsloth>=2024.12",
    "datasets>=3.0.0",
    "accelerate>=0.34.0",
    "scikit-learn>=1.5.0",
    "lm-eval>=0.4.0",
    "argilla>=2.0.0",
]
legacy-providers = [
    "anthropic>=0.37.0",
    "openai>=1.50.0",
    "google-cloud-aiplatform>=1.70.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/a1"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
